Move shared TypeScript into `@gradbench/common`
